+++ app/__init__.py
@@ -1,12 +1,14 @@
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+from flask_login import LoginManager
 
 from config import config
 
 
 db = SQLAlchemy()
 migrate = Migrate()
+login_manager = LoginManager()
 
 
 def create_app(config_name):
@@ -20,6 +22,7 @@ def create_app(config_name):
 
     db.init_app(app)
     migrate.init_app(app, db)
+    login_manager.init_app(app)
 
     with app.app_context():  # This makes config object available
         from .main import \